Moving carousel to homepage dir
