+++ docs/get-started/home.mdx
@@ -3,6 +3,6 @@ slug: /
 title: Glean Developer Platform
 ---
 
-import HomepageFeatures from '@site/src/components/HomepageFeatures'
+import Homepage from '@site/src/components/Homepage'
 
-<HomepageFeatures />
+<Homepage />
